Fixed concurrent modification problems
Changes:
- Removed ResourceHolder.hasResource method as not needed
- Fixed problem in UnitTests
- Fixed concurrent modification as People/Equipment disembark by holding
a private list copy
#463

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/equipment/EquipmentOwner.java
@@ -66,7 +66,7 @@ public interface EquipmentOwner extends ResourceHolder {
 	 * @param quantity
 	 * @return quantity that cannot be retrieved
 	 */
-	public double retrieveItemResource(int resource, int quantity);
+	public int retrieveItemResource(int resource, int quantity);
 	
 	/**
 	 * Gets the item resource stored
